added assistant and shared folders for components and moved corresponding components
