+++ src/components/assistant/ContractAssistant.tsx
